+++ CHANGELOG.md
@@ -23,6 +23,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Changed
 
+- [Library] Update border semantic tokens values ([#106](https://github.com/Orange-OpenSource/ouds-ios/issues/106))
 - [Showcase] Add fake components for demo and tokens tests
 - [Library] Remove spread value for elevation tokens
 - [Library] Remove paragraph spacing tokens for typography